Adding bash scripts

+++ bash/chrome-app/install.sh
@@ -6,11 +6,15 @@ LAYOUT_VERSION="24.05.13"
 LAYOUT_SEE="https://www.google.com/chrome/"
 
 LAYOUT_CODE() {
-  $IS_ROOT apt install $ARG_YES wget
-  wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-  $IS_ROOT dpkg -i google-chrome-stable_current_amd64.deb
-  $IS_ROOT apt-get $ARG_YES install -f
-  rm google-chrome-stable_current_amd64.deb
+  if ! command -v $LAYOUT_NAME &>/dev/null; then
+    $IS_ROOT apt install $ARG_YES wget
+    wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+    $IS_ROOT dpkg -i google-chrome-stable_current_amd64.deb
+    $IS_ROOT apt-get $ARG_YES install -f
+    rm google-chrome-stable_current_amd64.deb
+  else
+    echo $LAYOUT_APP is already
+  fi
 }
 
 # template
